requirements.txt
Django>=4.2.9,<5.0
djangorestframework==3.14.0

channels==4.0.0
channels-redis==4.1.0

celery==5.3.0
redis==5.0.1

mysqlclient==2.2.0
python-decouple==3.8

django-crispy-forms==2.0
crispy-bootstrap5==0.7

django-filter==23.3
django-cors-headers==4.2.0

geopy==2.4.0
folium==0.14.0

Pillow==10.1.0
requests==2.31.0
websockets==12.0
